requirements.txt
jira
ollama
tabulate
python-dotenv